requirement.txt
Django==1.8.6
Markdown==2.6.5
Pillow==3.0.0
PyYAML==3.11
argparse==1.2.1
django-filter==0.11.0
django-rest-swagger==0.3.4
djangorestframework==3.3.1
wsgiref==0.1.2
